refactor original code base to include teacher and person parent classes, write high fiveable module for shared behavior, add raise threshold constant in order to make performance rating method more dynamic
