+++ apprentice_teacher.rb
@@ -1,26 +1,17 @@
-require_relative 'high_fiveable'
 
-class ApprenticeTeacher
+require_relative 'teacher'
+
+class ApprenticeTeacher < Teacher
+
+  RATING_THRESHOLD = 80
 
   include HighFiveable
 
-  attr_reader :age, :salary, :phase, :target_raise
-  attr_accessor :name
+  attr_reader :salary, :target_raise
 
   def initialize(options={})
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-    @target_raise = 800
-    @phase = 3
-  end
-
-  # def offer_high_five
-  #   "High five!"
-  # end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
+    super
+    @target_raise = options.fetch(:target_raise, 800)
   end
 
   def teach_stuff
@@ -28,27 +19,6 @@ class ApprenticeTeacher
     response += "Listen, class, this is how everything works. "
     response += "*drops crazy knowledge bomb* "
     response += "... You're welcome."
-    response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
-
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 80
-      response = "Yay, I'm a great employee!"
-      receive_raise(@target_raise)
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
-    end
     response
   end
