+++ senior_teacher.rb
@@ -1,27 +1,19 @@
-require_relative 'high_fiveable'
 
-class SeniorTeacher
+require_relative 'teacher'
+
+class SeniorTeacher < Teacher
+
+  RATING_THRESHOLD = 90
 
   include HighFiveable
 
-  attr_reader :age, :salary, :phase, :performance_rating, :target_raise
-  attr_accessor :name
+  attr_reader :salary, :performance_rating, :target_raise
 
   def initialize(options={})
-    @phase = 3
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-    @target_raise = 1000
+    super
+    @target_raise = options.fetch(:target_raise, 1000)
   end
 
-  # def offer_high_five
-  #   "High five!"
-  # end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
 
   def teach_stuff
     response = ""
@@ -31,26 +23,6 @@ class SeniorTeacher
     response
   end
 
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
-
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 90
-      response = "Yay, I'm a great employee!"
-      receive_raise(@target_raise)
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
-    end
-    response
-  end
 
   def lead_training_session
     puts "Hey newbie!  Here are some common pitfalls.  Don't fall in them!"
