+++ student.rb
@@ -1,21 +1,14 @@
 require_relative 'high_fiveable'
+require_relative 'dbc_person'
 
-class Student
+class Student < DBCPerson
 
   include HighFiveable
 
-  attr_reader :age, :phase
-  attr_accessor :name
-
   def initialize(options = {})
-    @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
+   super
+   @phase = 1
   end
-
-  # def offer_high_five
-  #   "High five!"
-  # end
 
   def set_phase(num)
     response = ""
